requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
django-cors-headers
pillow